improved regnovo and related cnv.var.LocusSet: method to write set to file cnv.var.SampleData: minimal SampleData has CLASS= to avoid errors filesys.Segment: to string method seq.analysis.GATK_Genotyper: added threads to subset PlinkSeq changed method signatures and allow null vpop file PlinkSeqUtils: made things serializable, better reporting, ignore plinkseq gene definitions RegNovo: multiple improvements, WIP. VCFTally: serialized, individual level filtering per genotype, and perf improvements VCFTallyPSeq: multiple improvements SampleNGS: getter with hashset Project: method to load outlier hash from all samples VCFImporter: beginning to clean up and consolidate functionality VCFOps: serialized vpop, threaded extraction. added quick qc using plinkSeq VCOps: improvements for alt allele filtering, transmission VerifyNGS: removed unused imports BamQC: fixed error from the great deletion SeqError: re-worked for multi-threaded error iterations. some cleanup SeqQCValidation: improvements for multi-threading and speed FilterNGS: filters are now serializable

+++ src/seq/qc/BamQC.java
@@ -469,7 +469,7 @@ public class BamQC {
 	}
 
 	private static double computePercent(int top, int bottom) {
-		return bottom > 0 ? top / bottom : 0;
+		return bottom > 0 ? (double) top / bottom : 0;
 	}
 
 	private static BamQC initBamQC(String inputSamOrBamFile, String outputDir, FilterNGS filterNGS) {
